+++ client/src/components/Libraries/LibraryFolder/LibraryFolderPermissions/LibraryFolderPermissions.vue
@@ -57,7 +57,7 @@
 import Vue from "vue";
 import { getAppRoot } from "onload/loadConfig";
 import BootstrapVue from "bootstrap-vue";
-import { Services } from "./services";
+import { Services } from "components/Libraries/LibraryPermissions/services";
 import { Toast } from "ui/toast";
 import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import PermissionsInputField from "components/Libraries/LibraryPermissions/PermissionsInputField";
